requirements.txt
drf-yasg==1.20.0
amqp==5.0.6
arrow==1.1.1
asgiref==3.4.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
binaryornot==0.4.4
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-templated-mail==1.1.1
djangorestframework==3.12.4
djoser==2.1.0
drf-yasg2==1.19.4
flower==1.0.0
gunicorn==20.1.0
humanize==3.11.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jinja2-time==0.2.0
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.2
poyo==0.5.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
tornado==6.1
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0